slightly reorganize the code
Before this commit, there is a dev_apply_mic() function in each file that has functions related to the minimum image convention (periodic boundary conditions). Now two versions of this function are put into separate files. One of these two files is directly included into the files which need this device function. Because this is a device function, it does not compile separately without using the "-rdc=true" flag. However, it seems that the "-rdc=true" flag can heavily reduce the performance, which is unacceptable.

This is not a "good" way of coding, but I still haven't figured out a better way that can achieve the following two goals simultaneously:
1) avoid repetition (make the code as simple as possible)
2) keep the performance (make the code as fast as possible)