always calculate potential energy and pressure
Previously, I do not calculate the potential energy and pressure when the the heat conductivity (using any method) is to be calculated. This is not very good because one may want to check the energy and/or pressure in the production stage. 

This commit is only for the EAM potential. I will do the same for all other potentials.