Add phonon Python example & changed matlab to mirror get better match
Running the Jupyter notebook example will not give exact same results as the MATLAB version as the kpoints are different. This is a consequence of using ASE-based tools in Python. The MATLAB script has been updated to fix a bug, but also better match the Jupyter notebook. The 'dispersion.png' file still needs to be updated.